requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fake-useragent==1.4.0
Flask==3.0.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.36.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
murmurhash==1.0.10
numpy==1.26.2
packaging==23.2
pandas==2.1.3
preshed==3.0.9
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytrends==4.9.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
